requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.335
langchain-openai==0.0.2.post1
python-dotenv==1.0.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.8.6
httpx==0.25.1
tenacity==8.2.3
bcrypt==4.0.1